COMP: you guys should try 64bits...
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=7926
git-svn-id: http://svn.slicer.org/Slicer4/trunk@7926 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Base/GUI/vtkSlicerApplication.cxx
@@ -1715,7 +1715,7 @@ void vtkSlicerApplication::ArrayToString(vtkStringArray *array, std::string sep,
       res = res + sep;
       }
     res = res + s;
-    len += (s.size()+1);
+    len += (int)(s.size()+1);
     }
   strcpy(string, res.c_str());
 }
